chatbot_project/requirements.txt
google-generativeai>=0.3.2
python-dotenv>=1.0.0
requests>=2.31.0
Flask>=3.0.0
flask-cors>=4.0.0
